+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * [UIBULKED-4](https://issues.folio.org/browse/UIBULKED-4) Add Record identifier component to the Set Criteria pane
 * [UIBULKED-6](https://issues.folio.org/browse/UIBULKED-6) Add file drag and drop component
 * [UIBULKED-5](https://issues.folio.org/browse/UIBULKED-5) Add Apps with bulk edit capability accordion
+* [UIBULKED-8](https://issues.folio.org/browse/UIBULKED-8) Add Saved queries accordion